database connection, refactor, user stats
