+++ miejski.py
@@ -1,32 +1,39 @@
-import random
-import os
+from typing import List
 
-from discord.ext.commands import Context
+import requests
+from asyncpg import Record
+from bs4 import BeautifulSoup
+
 from utils import Utils
-from discord.ext import commands
 
 
-bot = commands.Bot(command_prefix='!')
+class Miejski:
+    @staticmethod
+    async def get_message() -> str:
+        url = 'https://www.miejski.pl/losuj'
+        r = requests.get(url, allow_redirects=True)
+        http = BeautifulSoup(r.text, 'html.parser')
+        redirected_url = http.find("link", {"rel": "canonical"})['href']
+        title = [x.get_text() for x in http.findAll("h1")]
+        definition = [x.get_text() for x in http.findAll("p")]
+        example = [x.get_text() for x in http.findAll("blockquote")]
+        rating = http.find("span", {"class": "rating"}).contents[0]
+        if len(example) > 0:
+            response = "**Słowo:** " + title[0] \
+                       + "\n**Ocena:** " + rating \
+                       + "\n**Definicja:** " + Utils.parse_html(definition[0]) \
+                       + "\n**Przykład:**" + Utils.parse_html(example[0]) \
+                       + "\n**URL**: " + redirected_url
+        else:
+            response = "**Słowo:** " + title[0] \
+                       + "\n**Ocena:** " + rating \
+                       + "\n**Definicja:** " + Utils.parse_html(definition[0]) \
+                       + "\n**URL**: " + redirected_url
+        return response
 
-
-@bot.event
-async def on_ready():
-    print('Logged in as')
-    print(bot.user.name)
-    print(bot.user.id)
-    print('------')
-    await bot.get_channel(271732666653474826).send("Jestem od teraz w nower wersji!")
-
-
-@bot.command()
-async def miejski(ctx: Context):
-    print('Recieved command !miejski from ' + ctx.author.name + ', processing...')
-    await ctx.send(await Utils.get_message())
-
-
-@bot.command(description='Pomoc w decyzjach')
-async def choose(ctx, *choices: str):
-    await ctx.send(random.choice(choices))
-
-
-bot.run(os.environ.get('DISCORD_TOKEN'))
+    @staticmethod
+    def get_stats(records: List[Record]):
+        result = "**STATYSTYKI**:\n"
+        for i, record in enumerate(records):
+            result += f'''**{i}. {record['USER_NAME']}**: {record['POINTS']} pkt. \n'''
+        return result
